final version

+++ src/modules/users/services/DeleteUserService/DeleteUserService.ts
@@ -1,5 +1,6 @@
 import { inject, injectable } from "tsyringe";
 
+import { AppError } from "../../../../errors/AppError";
 import { IUsersRepository } from "../../repositories/IUsersRepository";
 
 @injectable()
@@ -9,6 +10,10 @@ class DeleteUserService {
     private usersRepository: IUsersRepository
   ) {}
   async execute(_id: string): Promise<void> {
+    const user = await this.usersRepository.findUserById(_id);
+    if (!user) {
+      throw new AppError("User not found.", 404);
+    }
     await this.usersRepository.deleteUserById(_id);
   }
 }
